# src/main/java/com/example/droolsserver/controller/PorfiriaController.java
package com.example.droolsserver.controller;

import com.example.droolsserver.model.*;
import com.example.droolsserver.service.DroolsService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Map;
import java.util.HashMap;
import java.util.ArrayList;

@RestController
@RequestMapping("/api/porfiria")
@CrossOrigin(origins = "*")
public class PorfiriaController {

    @Autowired
    private DroolsService droolsService;

    @PostMapping("/evaluar")
    public ResponseEntity<Map<String, Object>> evaluarPaciente(
            @RequestBody Map<String, Object> request) {
        try {
            System.out.println("=== EVALUACIÓN DE PACIENTE ===");
            System.out.println("Request recibido: " + request);
            // Extraer datos del request
            Patient patient = new Patient();
            patient.setId((String) request.get("patientId"));
            patient.setFirstName((String) request.get("firstName"));
            patient.setLastName((String) request.get("lastName"));
            patient.setDni((String) request.get("dni"));
            patient.setAge((Integer) request.get("age"));
            patient.setGender((String) request.get("gender"));
            patient.setFamilyHistory((Boolean) request.get("familyHistory"));
            patient.setAlcoholConsumption((Boolean) request.get("alcoholConsumption"));
            patient.setFastingStatus((Boolean) request.get("fastingStatus"));

            // Crear respuestas del cuestionario
            @SuppressWarnings("unchecked")
            List<Map<String, Object>> responsesData = (List<Map<String, Object>>) request.get("responses");
            List<QuestionnaireResponse> responses = new ArrayList<>();
            
            if (responsesData != null) {
                for (Map<String, Object> responseData : responsesData) {
                    QuestionnaireResponse response = new QuestionnaireResponse();
                    response.setQuestionId((String) responseData.get("questionId"));
                    response.setAnswer((String) responseData.get("answer"));
                    response.setPatientId((String) responseData.get("patientId"));
                    responses.add(response);
                }
            }

            // Procesar con Drools
            System.out.println("Procesando con Drools...");
            List<Object> resultados = droolsService.obtenerResultadosCompletos(patient, responses);
            System.out.println("Resultados obtenidos: " + resultados.size());

            // Organizar resultados
            Map<String, Object> response = new HashMap<>();
            boolean tieneCuadro = false;
            boolean tieneDiagnostico = false;
            boolean tieneOrdenes = false;
            boolean tieneMedicamentos = false;

            for (Object resultado : resultados) {
                if (resultado instanceof CuadroClinico) {
                    response.put("cuadroClinico", resultado);
                    tieneCuadro = true;
                    System.out.println("Cuadro clínico encontrado: " + resultado);
                } else if (resultado instanceof DiagnosticoTemprano) {
                    response.put("diagnostico", resultado);
                    tieneDiagnostico = true;
                    System.out.println("Diagnóstico encontrado: " + resultado);
                } else if (resultado instanceof GenerarOrden) {
                    response.put("ordenes", resultado);
                    tieneOrdenes = true;
                    System.out.println("Órdenes encontradas: " + resultado);
                } else if (resultado instanceof InformarWeb) {
                    response.put("medicamentos", resultado);
                    tieneMedicamentos = true;
                    System.out.println("Medicamentos encontrados: " + resultado);
                }
            }

            // Completar estructura consistente con valores por defecto cuando no aplica
            if (!tieneDiagnostico) {
                DiagnosticoTemprano diagnosticoDefault = new DiagnosticoTemprano();
                diagnosticoDefault.setPatientId(patient.getId());
                diagnosticoDefault.setSintomaCutanea(false);
                diagnosticoDefault.setSintomaAguda(false);
                response.put("diagnostico", diagnosticoDefault);
            }

            if (!tieneOrdenes) {
                GenerarOrden ordenDefault = new GenerarOrden();
                ordenDefault.setPatientId(patient.getId());
                ordenDefault.setEstudios(false);
                response.put("ordenes", ordenDefault);
            }

            if (!tieneMedicamentos) {
                InformarWeb informarDefault = new InformarWeb();
                informarDefault.setPatientId(patient.getId());
                informarDefault.setMedicamentos(false);
                response.put("medicamentos", informarDefault);
            }

            System.out.println("Respuesta final: " + response);
            return ResponseEntity.ok(response);
        } catch (Exception e) {
            e.printStackTrace();
            return ResponseEntity.badRequest().build();
        }
    }

    @PostMapping("/cuadro-clinico")
    public ResponseEntity<CuadroClinico> obtenerCuadroClinico(
            @RequestBody Map<String, Object> request) {
        try {
            Patient patient = new Patient();
            patient.setId((String) request.get("patientId"));
            patient.setFirstName((String) request.get("firstName"));
            patient.setLastName((String) request.get("lastName"));
            patient.setDni((String) request.get("dni"));
            patient.setAge((Integer) request.get("age"));
            patient.setGender((String) request.get("gender"));

            @SuppressWarnings("unchecked")
            List<Map<String, Object>> responsesData = (List<Map<String, Object>>) request.get("responses");
            List<QuestionnaireResponse> responses = new ArrayList<>();
            
            if (responsesData != null) {
                for (Map<String, Object> responseData : responsesData) {
                    QuestionnaireResponse response = new QuestionnaireResponse();
                    response.setQuestionId((String) responseData.get("questionId"));
                    response.setAnswer((String) responseData.get("answer"));
                    response.setPatientId((String) responseData.get("patientId"));
                    responses.add(response);
                }
            }

            CuadroClinico cuadro = droolsService.procesarPaciente(patient, responses);
            return ResponseEntity.ok(cuadro);
        } catch (Exception e) {
            e.printStackTrace();
            return ResponseEntity.badRequest().build();
        }
    }

    @PostMapping("/diagnostico")
    public ResponseEntity<DiagnosticoTemprano> obtenerDiagnostico(
            @RequestBody Map<String, Object> request) {
        try {
            Patient patient = new Patient();
            patient.setId((String) request.get("patientId"));
            patient.setFirstName((String) request.get("firstName"));
            patient.setLastName((String) request.get("lastName"));
            patient.setDni((String) request.get("dni"));
            patient.setAge((Integer) request.get("age"));
            patient.setGender((String) request.get("gender"));

            @SuppressWarnings("unchecked")
            List<Map<String, Object>> responsesData = (List<Map<String, Object>>) request.get("responses");
            List<QuestionnaireResponse> responses = new ArrayList<>();
            
            if (responsesData != null) {
                for (Map<String, Object> responseData : responsesData) {
                    QuestionnaireResponse response = new QuestionnaireResponse();
                    response.setQuestionId((String) responseData.get("questionId"));
                    response.setAnswer((String) responseData.get("answer"));
                    response.setPatientId((String) responseData.get("patientId"));
                    responses.add(response);
                }
            }

            DiagnosticoTemprano diagnostico = droolsService.obtenerDiagnostico(patient, responses);
            return ResponseEntity.ok(diagnostico);
        } catch (Exception e) {
            e.printStackTrace();
            return ResponseEntity.badRequest().build();
        }
    }

    @PostMapping("/ordenes")
    public ResponseEntity<GenerarOrden> obtenerOrdenes(
            @RequestBody Map<String, Object> request) {
        try {
            Patient patient = new Patient();
            patient.setId((String) request.get("patientId"));
            patient.setFirstName((String) request.get("firstName"));
            patient.setLastName((String) request.get("lastName"));
            patient.setDni((String) request.get("dni"));
            patient.setAge((Integer) request.get("age"));
            patient.setGender((String) request.get("gender"));

            @SuppressWarnings("unchecked")
            List<Map<String, Object>> responsesData = (List<Map<String, Object>>) request.get("responses");
            List<QuestionnaireResponse> responses = new ArrayList<>();
            
            if (responsesData != null) {
                for (Map<String, Object> responseData : responsesData) {
                    QuestionnaireResponse response = new QuestionnaireResponse();
                    response.setQuestionId((String) responseData.get("questionId"));
                    response.setAnswer((String) responseData.get("answer"));
                    response.setPatientId((String) responseData.get("patientId"));
                    responses.add(response);
                }
            }

            GenerarOrden ordenes = droolsService.obtenerOrdenes(patient, responses);
            return ResponseEntity.ok(ordenes);
        } catch (Exception e) {
            e.printStackTrace();
            return ResponseEntity.badRequest().build();
        }
    }

    @GetMapping("/health")
    public ResponseEntity<String> health() {
        return ResponseEntity.ok("Sistema de Diagnóstico de Porfiria está funcionando correctamente");
    }

    @PostMapping("/test-simple")
    public ResponseEntity<Map<String, Object>> testSimple(@RequestBody Map<String, Object> request) {
        Map<String, Object> response = new HashMap<>();
        response.put("message", "Endpoint funcionando");
        response.put("receivedData", request);
        response.put("timestamp", new java.util.Date());
        return ResponseEntity.ok(response);
    }

}